A set of breaking changes. See the commit details.
- Introduced a factory pattern for `transformCommitForSub*()` functions. Adde a suffix `Factory` to their function names and lower cased file names.
- Commit notes will be printed before Features/Bug fixes/Other changes in the changelog.
- The `returnInvalidCommit` option for `transformCommitForSub*()` functions was maved from `context` object to a factory function (as `options`).
