+++ packages/ckeditor5-dev-env/lib/release-tools/tasks/generatechangelogforsubpackages.js
@@ -19,7 +19,7 @@ const executeOnPackages = require( '../utils/executeonpackages' );
 const getPackageJson = require( '../utils/getpackagejson' );
 const getSubPackagesPaths = require( '../utils/getsubpackagespaths' );
 const generateChangelogFromCommits = require( '../utils/generatechangelogfromcommits' );
-const transformCommitFunction = require( '../utils/transform-commit/transformcommitforsubpackage' );
+const transformCommitForSubPackageFactory = require( '../utils/transform-commit/transformcommitforsubpackagefactory' );
 
 /**
  * Generates the changelog for packages located in single repository.
@@ -80,6 +80,10 @@ module.exports = function generateChangelogForSubPackages( options ) {
 		log.info( '' );
 		log.info( chalk.bold.blue( `Generating changelog for "${ dependencyName }"...` ) );
 
+		const transformCommitFunction = transformCommitForSubPackageFactory( {
+			returnInvalidCommit: true
+		} );
+
 		return getNewReleaseType( transformCommitFunction, { tagName } )
 			.then( result => {
 				displayCommits( result.commits );
@@ -108,7 +112,7 @@ module.exports = function generateChangelogForSubPackages( options ) {
 					tagName,
 					isInternalRelease,
 					newTagName: packageJson.name + '@' + version,
-					transformCommit: transformCommitFunction
+					transformCommit: transformCommitForSubPackageFactory()
 				};
 
 				return generateChangelogFromCommits( changelogOptions )
