+++ packages/ckeditor5-dev-env/lib/release-tools/templates/footer.hbs
@@ -8,4 +8,3 @@
 
 {{/each}}
 {{/if}}
-
