+++ packages/ckeditor5-dev-env/lib/release-tools/templates/template.hbs
@@ -5,6 +5,7 @@ Internal changes only (updated dependencies, documentation, etc.).
 
 
 {{else}}
+{{> footer}}
 {{#each commitGroups}}
 ### {{title}}
 
@@ -20,5 +21,5 @@ Internal changes only (updated dependencies, documentation, etc.).
 Internal changes only (updated dependencies, documentation, etc.).
 
 {{/each}}
-{{> footer}}
+
 {{/if}}
